refactor(AES): new structure, split into different files
