--- src/aes.py
@@ -1,238 +1,0 @@
-from ast import List
-import copy
-
-
-sbox = [
-        0x63, 0x7c, 0x77, 0x7b, 0xf2, 0x6b, 0x6f, 0xc5, 0x30, 0x01, 0x67, 0x2b, 0xfe, 0xd7, 0xab, 0x76,
-        0xca, 0x82, 0xc9, 0x7d, 0xfa, 0x59, 0x47, 0xf0, 0xad, 0xd4, 0xa2, 0xaf, 0x9c, 0xa4, 0x72, 0xc0,
-        0xb7, 0xfd, 0x93, 0x26, 0x36, 0x3f, 0xf7, 0xcc, 0x34, 0xa5, 0xe5, 0xf1, 0x71, 0xd8, 0x31, 0x15,
-        0x04, 0xc7, 0x23, 0xc3, 0x18, 0x96, 0x05, 0x9a, 0x07, 0x12, 0x80, 0xe2, 0xeb, 0x27, 0xb2, 0x75,
-        0x09, 0x83, 0x2c, 0x1a, 0x1b, 0x6e, 0x5a, 0xa0, 0x52, 0x3b, 0xd6, 0xb3, 0x29, 0xe3, 0x2f, 0x84,
-        0x53, 0xd1, 0x00, 0xed, 0x20, 0xfc, 0xb1, 0x5b, 0x6a, 0xcb, 0xbe, 0x39, 0x4a, 0x4c, 0x58, 0xcf,
-        0xd0, 0xef, 0xaa, 0xfb, 0x43, 0x4d, 0x33, 0x85, 0x45, 0xf9, 0x02, 0x7f, 0x50, 0x3c, 0x9f, 0xa8,
-        0x51, 0xa3, 0x40, 0x8f, 0x92, 0x9d, 0x38, 0xf5, 0xbc, 0xb6, 0xda, 0x21, 0x10, 0xff, 0xf3, 0xd2,
-        0xcd, 0x0c, 0x13, 0xec, 0x5f, 0x97, 0x44, 0x17, 0xc4, 0xa7, 0x7e, 0x3d, 0x64, 0x5d, 0x19, 0x73,
-        0x60, 0x81, 0x4f, 0xdc, 0x22, 0x2a, 0x90, 0x88, 0x46, 0xee, 0xb8, 0x14, 0xde, 0x5e, 0x0b, 0xdb,
-        0xe0, 0x32, 0x3a, 0x0a, 0x49, 0x06, 0x24, 0x5c, 0xc2, 0xd3, 0xac, 0x62, 0x91, 0x95, 0xe4, 0x79,
-        0xe7, 0xc8, 0x37, 0x6d, 0x8d, 0xd5, 0x4e, 0xa9, 0x6c, 0x56, 0xf4, 0xea, 0x65, 0x7a, 0xae, 0x08,
-        0xba, 0x78, 0x25, 0x2e, 0x1c, 0xa6, 0xb4, 0xc6, 0xe8, 0xdd, 0x74, 0x1f, 0x4b, 0xbd, 0x8b, 0x8a,
-        0x70, 0x3e, 0xb5, 0x66, 0x48, 0x03, 0xf6, 0x0e, 0x61, 0x35, 0x57, 0xb9, 0x86, 0xc1, 0x1d, 0x9e,
-        0xe1, 0xf8, 0x98, 0x11, 0x69, 0xd9, 0x8e, 0x94, 0x9b, 0x1e, 0x87, 0xe9, 0xce, 0x55, 0x28, 0xdf,
-        0x8c, 0xa1, 0x89, 0x0d, 0xbf, 0xe6, 0x42, 0x68, 0x41, 0x99, 0x2d, 0x0f, 0xb0, 0x54, 0xbb, 0x16
-]
-
-invSBOX = [
-        0x52, 0x09, 0x6a, 0xd5, 0x30, 0x36, 0xa5, 0x38, 0xbf, 0x40, 0xa3, 0x9e, 0x81, 0xf3, 0xd7, 0xfb,
-        0x7c, 0xe3, 0x39, 0x82, 0x9b, 0x2f, 0xff, 0x87, 0x34, 0x8e, 0x43, 0x44, 0xc4, 0xde, 0xe9, 0xcb,
-        0x54, 0x7b, 0x94, 0x32, 0xa6, 0xc2, 0x23, 0x3d, 0xee, 0x4c, 0x95, 0x0b, 0x42, 0xfa, 0xc3, 0x4e,
-        0x08, 0x2e, 0xa1, 0x66, 0x28, 0xd9, 0x24, 0xb2, 0x76, 0x5b, 0xa2, 0x49, 0x6d, 0x8b, 0xd1, 0x25,
-        0x72, 0xf8, 0xf6, 0x64, 0x86, 0x68, 0x98, 0x16, 0xd4, 0xa4, 0x5c, 0xcc, 0x5d, 0x65, 0xb6, 0x92,
-        0x6c, 0x70, 0x48, 0x50, 0xfd, 0xed, 0xb9, 0xda, 0x5e, 0x15, 0x46, 0x57, 0xa7, 0x8d, 0x9d, 0x84,
-        0x90, 0xd8, 0xab, 0x00, 0x8c, 0xbc, 0xd3, 0x0a, 0xf7, 0xe4, 0x58, 0x05, 0xb8, 0xb3, 0x45, 0x06,
-        0xd0, 0x2c, 0x1e, 0x8f, 0xca, 0x3f, 0x0f, 0x02, 0xc1, 0xaf, 0xbd, 0x03, 0x01, 0x13, 0x8a, 0x6b,
-        0x3a, 0x91, 0x11, 0x41, 0x4f, 0x67, 0xdc, 0xea, 0x97, 0xf2, 0xcf, 0xce, 0xf0, 0xb4, 0xe6, 0x73,
-        0x96, 0xac, 0x74, 0x22, 0xe7, 0xad, 0x35, 0x85, 0xe2, 0xf9, 0x37, 0xe8, 0x1c, 0x75, 0xdf, 0x6e,
-        0x47, 0xf1, 0x1a, 0x71, 0x1d, 0x29, 0xc5, 0x89, 0x6f, 0xb7, 0x62, 0x0e, 0xaa, 0x18, 0xbe, 0x1b,
-        0xfc, 0x56, 0x3e, 0x4b, 0xc6, 0xd2, 0x79, 0x20, 0x9a, 0xdb, 0xc0, 0xfe, 0x78, 0xcd, 0x5a, 0xf4,
-        0x1f, 0xdd, 0xa8, 0x33, 0x88, 0x07, 0xc7, 0x31, 0xb1, 0x12, 0x10, 0x59, 0x27, 0x80, 0xec, 0x5f,
-        0x60, 0x51, 0x7f, 0xa9, 0x19, 0xb5, 0x4a, 0x0d, 0x2d, 0xe5, 0x7a, 0x9f, 0x93, 0xc9, 0x9c, 0xef,
-        0xa0, 0xe0, 0x3b, 0x4d, 0xae, 0x2a, 0xf5, 0xb0, 0xc8, 0xeb, 0xbb, 0x3c, 0x83, 0x53, 0x99, 0x61,
-        0x17, 0x2b, 0x04, 0x7e, 0xba, 0x77, 0xd6, 0x26, 0xe1, 0x69, 0x14, 0x63, 0x55, 0x21, 0x0c, 0x7d
-]
-
-rcon = [
-    0x01, 0x00, 0x00, 0x00, 0x02, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x00, 0x1b, 0x00, 0x00, 0x00, 0x36, 0x00, 0x00, 0x00
-]
-
-# key schedule
-
-def SubBytes(currentCol):
-    return [sbox[i] for i in currentCol]
-
-def RotWord(word):
-    return word[1:] + word[:1]
-
-def InvSubBytes(word):
-    return [invSBOX[i] for i in word]
-
-def ShiftRows(word):
-
-    new = list(word)
-    new[1] = word[5]
-    new[5] = word[9]
-    new[9] = word[13]
-    new[13] = word[1]
-
-    new[2] = word[10]
-    new[6] = word[14]
-    new[10] = word[2]
-    new[14] = word[6]
-
-    new[3] = word[15]
-    new[7] = word[3]
-    new[11] = word[7]
-    new[15] = word[11]
-
-    return new
-
-def InvShiftRows(word):
-
-    new = list(word)
-    new[1] = word[13]
-    new[5] = word[1]
-    new[9] = word[5]
-    new[13] = word[9]
-
-    new[2] = word[10]
-    new[6] = word[14]
-    new[10] = word[2]
-    new[14] = word[6]
-
-    new[3] = word[7]
-    new[7] = word[11]
-    new[11] = word[15]
-    new[15] = word[3]
-
-    return new
-
-def addRoundKey(word1,word2): #addroundkey --!!
-    return [word1[i] ^ word2[i] for i in range(0, len(word1))]
-
-def printLine(line):
-    print(f"{line[0]:x} {line[0+1]:x} {line[0+2]:x} {line[0+3]:x}")
-
-def printTab(keymatrix):
-    for i in range(0, len(keymatrix), 4):
-            print(f"{keymatrix[i]:x} {keymatrix[i+1]:x} {keymatrix[i+2]:x} {keymatrix[i+3]:x}")
-
-def keyExpansion(key : List):
-    global rcon
-
-    for i in range(0,10):
-        for i in range(0,4):
-            newline = key[len(key) - 4: len(key) + 1]
-            if (i == 0):
-                newline = RotWord(newline)
-                newline = SubBytes(newline)
-                rConCol = rcon[0:4]
-                rcon = rcon[4:]
-                newline = addRoundKey(newline, rConCol)
-            wL4 = key[len(key) - 16: len(key) - 12]
-            newline = addRoundKey(newline, wL4)
-            key += newline
-    return key
-
-
-def gal(a, b):
-    res = 0
-    wrk = 0
-    interv = 256
-    rangeToSeek = 8
-
-    for i in range(rangeToSeek):
-        if b & 1 == 1: res = res ^ a
-        wrk = a & 0x80
-        a <<= 1
-        if wrk == 0x80: a = a ^ 0x1b
-        b >>= 1
-    return res % interv
-
-def mixColmn(colmn):
-    tmp = list(colmn)
-    colmn[0] = gal(tmp[0],2) ^ tmp[3] ^ tmp[2] ^ gal(tmp[1],3)
-    colmn[1] = gal(tmp[1],2) ^ tmp[0] ^ tmp[3] ^ gal(tmp[2],3)
-    colmn[2] = gal(tmp[2],2) ^ tmp[1] ^ tmp[0] ^ gal(tmp[3],3)
-    colmn[3] = gal(tmp[3],2) ^ tmp[2] ^ tmp[1] ^ gal(tmp[0],3)
-
-def invMixColmn(colmn): #tmp
-    tmp = list(colmn)
-    colmn[0] = gal(tmp[0], 14) ^ gal(tmp[3], 9) ^ gal(tmp[2], 13) ^ gal(tmp[1], 11)
-    colmn[1] = gal(tmp[1], 14) ^ gal(tmp[0], 9) ^ gal(tmp[3], 13) ^ gal(tmp[2], 11)
-    colmn[2] = gal(tmp[2], 14) ^ gal(tmp[1], 9) ^ gal(tmp[0], 13) ^ gal(tmp[3], 11)
-    colmn[3] = gal(tmp[3], 14) ^ gal(tmp[2], 9) ^ gal(tmp[1], 13) ^ gal(tmp[0], 11)
-
-
-def mixIt(message, mixFunc):
-    for i in range(4):
-        line = message[i*4:i*4+4]
-        mixFunc(line)
-        message[i*4:i*4+4] = line
-
-def rev_expandedkey(keyExpanded):
-    keys = [keyExpanded[i:i+16] for i in range(0, len(keyExpanded), 16)]
-    keys = keys[::-1]
-    res = []
-    for i in range(len(keys)):
-        for k in keys[i]:
-            res.append(k)
-    return res
-
-
-def aes_decrypt(message, keyExpanded):
-    keyExpanded = rev_expandedkey(keyExpanded)
-    message = addRoundKey(message, keyExpanded[:16])
-    keyExpanded = keyExpanded[16:]
-
-    message = InvShiftRows(message)
-    message = InvSubBytes(message)
-
-    for i in range(0,9):
-        message = addRoundKey(message, keyExpanded[0:16])
-        keyExpanded = keyExpanded[16:]
-        mixIt(message, invMixColmn)
-        message = InvShiftRows(message)
-        message = InvSubBytes(message)
-
-    message = addRoundKey(message, keyExpanded[0:16])
-    keyExpanded = keyExpanded[16:]
-
-    return message
-
-
-def aes_crypt(message, keyExpanded):
-
-    message = addRoundKey(message, keyExpanded[0:16])
-    keyExpanded = keyExpanded[16:]
-
-    for i in range(0, 9):
-        message = SubBytes(message)
-
-        message = ShiftRows(message)
-
-        mixIt(message, mixColmn)
-
-        message = addRoundKey(message, keyExpanded[0:16])
-        keyExpanded = keyExpanded[16:]
-
-    message = SubBytes(message)
-    message = ShiftRows(message)
-    message = addRoundKey(message, keyExpanded[0:16])
-    keyExpanded = keyExpanded[16:]
-
-    return message
-
-
-key = [
-        0x2b, 0x7e, 0x15,0x16, 0x28,0xae, 0xd2, 0xa6, 0xab, 0xf7, 0x15, 0x88, 0x09, 0xcf, 0x4f, 0x3c
-    ]
-
-message = [
-        0x32, 0x43, 0xf6, 0xa8, 0x88, 0x5a, 0x30, 0x8d, 0x31, 0x31, 0x98, 0xa2, 0xe0, 0x37, 0x07, 0x34
-    ]
-
-
-
-keyExpansion(key)
-keyExpandedSave = list(key)
-
-print("tab natural")
-printTab(message)
-
-message_cyphered = aes_crypt(message, key)
-
-print("tab cyphered")
-printTab(message_cyphered)
-
-print("tab uncyphered")
-message_decyphered = aes_decrypt(message_cyphered, keyExpandedSave)
-printTab(message_decyphered)
